+++ llm_gent/core/tools/builtin/__init__.py
@@ -1,13 +1,17 @@
 """Built-in tools for common operations."""
 
-from .brave_search import BraveSearchBackend
 from .complete import CompleteTaskTool
 from .file import FileReadTool, FileWriteTool
 from .http import HTTPFetchTool
 from .learn import RecallTool, RememberTool
 from .shell import ShellTool
-from .web_fetch import WebFetchTool
-from .web_search import WebSearchTool
+from .web import (
+    BraveSearchBackend,
+    WebFetchTool,
+    WebSearchBackend,
+    WebSearchBackendFactory,
+    WebSearchTool,
+)
 
 
 __all__ = [
@@ -20,5 +24,7 @@ __all__ = [
     "RememberTool",
     "ShellTool",
     "WebFetchTool",
+    "WebSearchBackend",
+    "WebSearchBackendFactory",
     "WebSearchTool",
 ]